requirements.txt
annotated-types==0.6.0
anyio==3.7.1
assemblyai==0.19.0
backports.zoneinfo==0.2.1
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
distro==1.8.0
exceptiongroup==1.1.3
google-api-core==2.14.0
google-auth==2.23.4
google-cloud-texttospeech==2.14.2
googleapis-common-protos==1.61.0
grpcio==1.59.2
grpcio-status==1.59.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
openai==1.2.3
proto-plus==1.22.3
protobuf==4.25.0
pvporcupine==3.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.14
pydantic==2.4.2
pydantic-core==2.10.1
pydub==0.25.1
PySDL2==0.9.16
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
sniffio==1.3.0
tqdm==4.66.1
typing-extensions==4.8.0
tzlocal==5.2
urllib3==2.0.7
webrtcvad==2.0.10
websockets==12.0
